requirements.txt
requests==2.32.3
player-2.1.0

